+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/model/NrcHorarioPK.java
@@ -8,19 +8,19 @@ import javax.persistence.Embeddable;
 public class NrcHorarioPK implements Serializable {
 
     @Column(name = "cod_nrc", nullable = false)
-    private short codNrc;
+    private Integer codNrc;
 
     @Column(name = "cod_periodo", nullable = false)
-    private short codPeriodo;
+    private Integer codPeriodo;
 
     @Column(name = "cod_departamento", nullable = false)
-    private int codDepartamento;
+    private Integer codDepartamento;
 
     @Column(name = "cod_materia", nullable = false)
-    private int codMateria;
+    private Integer codMateria;
 
     @Column(name = "cod_aula", nullable = false)
-    private int codAula;
+    private Integer codAula;
 
     @Column(name = "dia_semana", nullable = false, length = 25)
     private String diaSemana;
@@ -28,7 +28,7 @@ public class NrcHorarioPK implements Serializable {
     public NrcHorarioPK() {
     }
 
-    public NrcHorarioPK(short codNrc, short codPeriodo, int codDepartamento, int codMateria, int codAula, String diaSemana) {
+    public NrcHorarioPK(Integer codNrc, Integer codPeriodo, Integer codDepartamento, Integer codMateria, Integer codAula, String diaSemana) {
         this.codNrc = codNrc;
         this.codPeriodo = codPeriodo;
         this.codDepartamento = codDepartamento;
@@ -37,43 +37,43 @@ public class NrcHorarioPK implements Serializable {
         this.diaSemana = diaSemana;
     }
 
-    public short getCodNrc() {
+    public Integer getCodNrc() {
         return codNrc;
     }
 
-    public void setCodNrc(short codNrc) {
+    public void setCodNrc(Integer codNrc) {
         this.codNrc = codNrc;
     }
 
-    public short getCodPeriodo() {
+    public Integer getCodPeriodo() {
         return codPeriodo;
     }
 
-    public void setCodPeriodo(short codPeriodo) {
+    public void setCodPeriodo(Integer codPeriodo) {
         this.codPeriodo = codPeriodo;
     }
 
-    public int getCodDepartamento() {
+    public Integer getCodDepartamento() {
         return codDepartamento;
     }
 
-    public void setCodDepartamento(int codDepartamento) {
+    public void setCodDepartamento(Integer codDepartamento) {
         this.codDepartamento = codDepartamento;
     }
 
-    public int getCodMateria() {
+    public Integer getCodMateria() {
         return codMateria;
     }
 
-    public void setCodMateria(int codMateria) {
+    public void setCodMateria(Integer codMateria) {
         this.codMateria = codMateria;
     }
 
-    public int getCodAula() {
+    public Integer getCodAula() {
         return codAula;
     }
 
-    public void setCodAula(int codAula) {
+    public void setCodAula(Integer codAula) {
         this.codAula = codAula;
     }
 
@@ -88,11 +88,11 @@ public class NrcHorarioPK implements Serializable {
     @Override
     public int hashCode() {
         int hash = 0;
-        hash += (int) codNrc;
-        hash += (int) codPeriodo;
-        hash += (int) codDepartamento;
-        hash += (int) codMateria;
-        hash += (int) codAula;
+        hash += (Integer) codNrc;
+        hash += (Integer) codPeriodo;
+        hash += (Integer) codDepartamento;
+        hash += (Integer) codMateria;
+        hash += (Integer) codAula;
         hash += (diaSemana != null ? diaSemana.hashCode() : 0);
         return hash;
     }
