+++ src/main/java/ec/edu/espe/arquitectura/escolastico/orgfisico/model/Institucion.java
@@ -18,7 +18,7 @@ public class Institucion implements Serializable {
     private static final long serialVersionUID = 120L;
     @Id
     @Column(name = "cod_institucion", nullable = false)
-    private Short codInstitucion;
+    private Integer codInstitucion;
 
     @Column(name = "ruc", nullable = false, length = 13)
     private String ruc;
@@ -43,7 +43,7 @@ public class Institucion implements Serializable {
     private String audIp;
 
     @Column(name = "version", nullable = false)
-    private int version;
+    private Integer version;
 
     @OneToMany(mappedBy = "institucion")
     private List<Sede> sedes;
@@ -51,15 +51,15 @@ public class Institucion implements Serializable {
     public Institucion() {
     }
 
-    public Institucion(Short codInstitucion) {
+    public Institucion(Integer codInstitucion) {
         this.codInstitucion = codInstitucion;
     }
 
-    public Short getCodInstitucion() {
+    public Integer getCodInstitucion() {
         return codInstitucion;
     }
 
-    public void setCodInstitucion(Short codInstitucion) {
+    public void setCodInstitucion(Integer codInstitucion) {
         this.codInstitucion = codInstitucion;
     }
 
@@ -119,11 +119,11 @@ public class Institucion implements Serializable {
         this.audIp = audIp;
     }
 
-    public int getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(int version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
